added gzip compression
